+++ tool_packages/unique_internal_search/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.8] - 2025-11-20
+- Update toolkit version
+
+## [1.2.7] - 2025-11-20
+- Include message log messages
+
+## [1.2.6] - 2025-11-20
+- Fix bug of handling properly uploaded files that are expired1
+
+## [1.2.5] - 2025-11-20
+- Bump tiktoken to 0.12.0
+
 ## [1.2.4] - 2025-11-17
 - Fix bug in search method when `exclude_uploaded_files` is True