+++ unique_toolkit/CHANGELOG.md
@@ -5,6 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.29.0] - 2025-11-21
+- Add option to force include references in sub agent responses even if unused by main agent response.
+
+## [1.28.9] - 2025-11-21
+- Remove `knolwedge_base_service` from DocXGeneratorService
+
+## [1.28.8] - 2025-11-20
+- Add query params to api operation
+- Add query params to endpoint builder
+
+## [1.28.7] - 2025-11-20
+- Adding Message Step Logger Class to the agentic tools.
+
+## [1.28.6] - 2025-11-20
+- Adding tests for message role filtering in chat functions
+
+## [1.28.5] - 2025-11-20
+- Including `expired_at` parameter in content schema
+- Including chatRole `SYSTEM` into chat schema
+
+## [1.28.4] - 2025-11-20
+- Bump tiktoken to 0.12.0
+
+## [1.28.3] - 2025-11-20
+- Add batch upload of files to knowledgebase
+- Add create folders utilities
+
 ## [1.28.2] - 2025-11-20
 - Adding model `litellm:gemini-3-pro-preview` to `language_model/info.py`